Image upload, slug, user permission, customized filter

+++ Blog/posts/views.py
@@ -1,4 +1,5 @@
-from django.http import HttpResponseRedirect
+from urllib import quote_plus
+from django.http import HttpResponseRedirect, Http404
 from django.shortcuts import render, get_object_or_404, redirect
 from django.core.paginator import Paginator, EmptyPage, PageNotAnInteger
 from django.shortcuts import render
@@ -10,10 +11,14 @@ from .forms import PostForm
 
 
 def post_create(request):
-    form = PostForm(request.POST or None)  # or None : Remember this!
+    if not request.user.is_staff or not request.user.is_superuser:
+        raise Http404
+    form = PostForm(request.POST or None, request.FILES or None)  # or None : Remember this!
     if request.method == 'POST':
         if form.is_valid():
-            obj = form.save()
+            obj = form.save(commit=False)
+            obj.user = request.user
+            obj.save()
             messages.success(request, '<a href="' + obj.get_absolute_url() + '">' + "Successfully created" + "</a>", extra_tags="html_safe")
             return HttpResponseRedirect(obj.get_absolute_url())
         else:
@@ -22,11 +27,13 @@ def post_create(request):
         return render(request, "posts/post_form.html", context={'form': form})
 
 
-def post_detail(request, id):
-    obj = get_object_or_404(Post, id=id)
+def post_detail(request, slug):
+    obj = get_object_or_404(Post, slug=slug)
+    share_string = quote_plus(obj.content)
     context = {
         'title': obj.title,
-        'obj': obj
+        'obj': obj,
+        'share_string': share_string
     }
     return render(request, "posts/detail.html", context)
 
@@ -50,9 +57,11 @@ def post_list(request):
     }
     return render(request, "posts/post_list.html", context)
 
-def post_update(request, id):
-    obj = get_object_or_404(Post, id=id)
-    form = PostForm(request.POST or None, instance=obj)
+def post_update(request, slug):
+    if not request.user.is_staff or not request.user.is_superuser:
+        raise Http404
+    obj = get_object_or_404(Post, slug=slug)
+    form = PostForm(request.POST or None, request.FILES or None, instance=obj)
     if form.is_valid():
         instance = form.save(commit=False)
         instance.save()
@@ -66,8 +75,10 @@ def post_update(request, id):
     }
     return render(request,"posts/post_form.html",context)
 
-def post_delete(request, id):
-    obj = get_object_or_404(Post, id=id)
+def post_delete(request, slug):
+    if not request.user.is_staff or not request.user.is_superuser:
+        raise Http404
+    obj = get_object_or_404(Post, slug=slug)
     obj.delete()
     messages.success(request,"Post deleted")
     return redirect("posts:list")
